refactor(state): move RocksDB-specific initialization to a new module
