+++ zebra-state/src/lib.rs
@@ -12,6 +12,9 @@
 #![doc(html_logo_url = "https://www.zfnd.org/images/zebra-icon.png")]
 #![doc(html_root_url = "https://doc.zebra.zfnd.org/zebra_state")]
 
+#[macro_use]
+extern crate tracing;
+
 #[cfg(any(test, feature = "proptest-impl"))]
 mod arbitrary;
 mod config;
